// src/match-engine/types.ts
export type MovementState =
  | 'IDLE'
  | 'POSITIONING'
  | 'MOVING_TO_BALL'
  | 'SUPPORT'
  | 'DEFENDING'
  | 'ATTACKING'
  | 'ON_BALL'
  | 'RECEIVING'
  | 'PRESSING';